src/main/persistence/JsonWriter.java
package persistence;


import model.QuantumCircuit;
import org.json.JSONObject;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.PrintWriter;

// Represents a writer that writes a JSON representation of a QuantumCircuit to file.
// Citation: code obtained from JsonWriter class in JsonSerializationDemo
//           URL: https://github.students.cs.ubc.ca/CPSC210/JsonSerializationDemo.git
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // Citation: Taken from JsonSerializationDemo
    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // Citation: Taken from JsonSerializationDemo
    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    // be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // Citation: Taken from JsonSerializationDemo
    // MODIFIES: this
    // EFFECTS: writes JSON representation of QuantumCircuit to file
    public void write(QuantumCircuit qc) {
        JSONObject json = qc.toJson();
        saveToFile(json.toString(TAB));
    }

    // Citation: Taken from JsonSerializationDemo
    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // Citation: Taken from JsonSerializationDemo
    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }

}
